added degrade_juiciness to model

+++ db/migrate/20140705002204_create_secrets.rb
@@ -2,7 +2,7 @@ class CreateSecrets < ActiveRecord::Migration
   def change
     create_table :secrets do |t|
       t.string :content
-      t.integer :juiciness, default: 10
+      t.float :juiciness, default: 10
 
       t.timestamps
     end
